Split out createRule utility
